requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-filter>=24.3,<25.0
psycopg[binary]>=3.2,<3.3
python-dotenv>=1.0,<2.0
